+++ gateway/src/main/java/ru/practicum/shareit/request/ItemRequestController.java
@@ -23,33 +23,38 @@ import ru.practicum.shareit.request.dto.ItemRequestDto;
 @RequestMapping(path = "/requests")
 @AllArgsConstructor
 public class ItemRequestController {
-    private final ItemRequestClient client;
+	private final ItemRequestClient client;
 
-    private static final String HEADER_USER_ID = "X-Sharer-User-Id";
+	private static final String HEADER_USER_ID = "X-Sharer-User-Id";
 
-    @GetMapping
-    public ResponseEntity<Object> getAllRequests(@RequestHeader(HEADER_USER_ID) Long userId) {
-        return client.getAllRequests(userId);
-    }
+	@GetMapping
+	public ResponseEntity<Object> getAllRequests(
+			@RequestHeader(HEADER_USER_ID) long userId) {
+		return client.getAllRequests(userId);
+	}
 
-    @GetMapping("/{requestId}")
-    public ResponseEntity<Object> getRequest(@RequestHeader(HEADER_USER_ID) Long userId, @PathVariable Long requestId) {
-        return client.getRequest(userId.longValue(), requestId.longValue());
-    }
+	@GetMapping("/{requestId}")
+	public ResponseEntity<Object> getRequest(
+			@RequestHeader(HEADER_USER_ID) long userId,
+			@PathVariable long requestId) {
+		return client.getRequest(userId, requestId);
+	}
 
-    @GetMapping("/all")
-    public ResponseEntity<Object> getAllUsersRequests(@RequestHeader(HEADER_USER_ID) Long userId,
-            @RequestParam(required = false, defaultValue = "0") Integer from,
-            @RequestParam(required = false, defaultValue = "10") Integer size) {
-        if (from.longValue() < 0) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
-        }
-        return client.getAllUsersRequests(userId.longValue(), from.intValue(), size.intValue());
-    }
+	@GetMapping("/all")
+	public ResponseEntity<Object> getAllUsersRequests(
+			@RequestHeader(HEADER_USER_ID) long userId,
+			@RequestParam(required = false, defaultValue = "0") int from,
+			@RequestParam(required = false, defaultValue = "10") int size) {
+		if (from < 0) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+		}
+		return client.getAllUsersRequests(userId, from, size);
+	}
 
-    @PostMapping
-    public ResponseEntity<Object> createRequest(@RequestHeader(HEADER_USER_ID) Long userId,
-            @RequestBody @Valid ItemRequestDto itemRequestDto) {
-        return client.createRequest(userId.longValue(), itemRequestDto);
-    }
+	@PostMapping
+	public ResponseEntity<Object> createRequest(
+			@RequestHeader(HEADER_USER_ID) long userId,
+			@RequestBody @Valid ItemRequestDto itemRequestDto) {
+		return client.createRequest(userId, itemRequestDto);
+	}
 }
